Draw mainTexture property as a first-draft texture preview
The goal of this feature is a nice texture preview, which allows users to select the material/texture from the mesh, pick image color channels to display and set the alpha value.

This commit shows the first prototype which draws a simple texture into the UV window. While testing this, it was discovered that the UVs are potentially flipped and that the texture tiling and offset values are not correctly handled yet.